backend/pyproject.toml
[tool.poetry]
name = "handwriteai-backend"
version = "0.1.0"
description = "Backend for HandwriteAI - Text to Handwriting Conversion"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.4.2"
python-multipart = "^0.0.6"
torch = "^2.2.0"
numpy = "^1.26.0"
pillow = "^10.0.0"
redis = "^5.0.1"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
supabase = "^2.3.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 